refactor(otel): split bridge files, bump deps, wire grafana datasource
- Split main.go (770 lines) into main/config/engine/mux/stream/writer/
  checkpoint/util.go (each 60-145 lines). Cut redundant comments.
- Bump Go 1.23 -> 1.26, clickhouse 24.8 -> 26.3, connectrpc.com/connect
  1.18 -> 1.19, golang.org/x/net 0.33 -> 0.54, kurtosis/api/golang
  1.5 -> 1.18.2, google.golang.org/protobuf 1.34 -> 1.36.
- When the `otel` additional_service is enabled alongside `grafana`,
  install the grafana-clickhouse-datasource plugin and register the
  in-enclave ClickHouse as a datasource pointing at otel.otel_logs.
