+++ src/otel/bridge/go.mod
@@ -1,18 +1,17 @@
 module github.com/ethpandaops/ethereum-package/otel-bridge
 
-go 1.23
+go 1.26.0
 
 require (
-	connectrpc.com/connect v1.18.1
-	github.com/kurtosis-tech/kurtosis/api/golang v1.5.0
-	golang.org/x/net v0.33.0
-	google.golang.org/protobuf v1.34.2
+	connectrpc.com/connect v1.19.2
+	github.com/kurtosis-tech/kurtosis/api/golang v1.18.2
+	golang.org/x/net v0.54.0
+	google.golang.org/protobuf v1.36.11
 )
 
 require (
-	github.com/golang/protobuf v1.5.3 // indirect
-	golang.org/x/sys v0.28.0 // indirect
-	golang.org/x/text v0.21.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20230803162519-f966b187b2e5 // indirect
-	google.golang.org/grpc v1.57.1 // indirect
+	golang.org/x/sys v0.44.0 // indirect
+	golang.org/x/text v0.37.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20260120221211-b8f7ae30c516 // indirect
+	google.golang.org/grpc v1.80.0 // indirect
 )
