+++ static_files/grafana-config/templates/datasource.yml.tmpl
@@ -27,3 +27,22 @@ datasources:
         spanEndTimeShift: '1h'
         filterByTraceID: true
 {{- end }}
+{{- if .ClickHouseHost }}
+  - name: ClickHouse (otel)
+    type: grafana-clickhouse-datasource
+    access: proxy
+    orgId: 1
+    isDefault: false
+    editable: true
+    jsonData:
+      host: {{ .ClickHouseHost }}
+      port: {{ .ClickHousePort }}
+      protocol: http
+      secure: false
+      defaultDatabase: otel
+      logs:
+        defaultTable: otel_logs
+        timeColumn: Timestamp
+        levelColumn: SeverityText
+        messageColumn: Body
+{{- end }}
